+++ src/Model/Doc/Category.ts
@@ -1,8 +1,9 @@
-/// レコードカテゴリデータ。
+/// レコードのカテゴリ。
 class Category {
   id: number = 0; ///< Id。
   name: string = ""; ///< カテゴリ名。
-  parent: number = 0; ///< 親カテゴリのId。0 ならルート。
+  parent: Category | null = null; ///< 親 Category の参照。null ならルート。
+  childs: Category[] = []; ///< 子 Category の配列。
 };
 
 export default Category;
