+++ src/Model/Doc/CategoryKind.ts
