+++ src/Model/Doc/Data/Category.ts
@@ -1,0 +1,8 @@
+/// レコードカテゴリデータ。
+class Category {
+  id: number = 0; ///< Id。
+  name: string = ""; ///< カテゴリ名。
+  parent: number = 0; ///< 親カテゴリのId。0 ならルート。
+};
+
+export default Category;
