+++ src/Model/Doc/Data/Record.ts
@@ -1,0 +1,7 @@
+/// 入出金レコード共通クラスデータ版。
+class Record {
+  date: string = ""; ///< 入出金発生日付。(YYYY-MM-DD形式)
+  memo: string = ""; ///< メモ。
+};
+
+export default Record;
