+++ src/Model/Doc/RecordKind.ts
